🏗♻️ Refactor / clean up the `watch` logic used for extensions (#22659)
